requirements-dev.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
Pillow==8.1.0
python-decouple==3.3
dj-database-url==0.5.0
dj-static==0.0.6